Add Hiera key store backend API and step method
This commit adds a new step method for Pipeline scripts to set key/value
pairs for specific environments.  The method invocation looks like this:

puppetHiera environment: 'production', key: 'key', value: 'value'

A HTTP /hiera/lookup endpoint is added to the Jenkins server that
accepts form data with a key *json* where the value of that key is the
JSON request. This endpoint is not authenticated in any way so no
secrets should be stored in it.

An example curl looks like this:

curl -X POST -d 'json={"environment": "production","key":
"class::parameter"}' http://jenkins.example.com:8080/jenkins/hiera/lookup

An accomponying Hiera backend will enable Hiera to use this API to
lookup key/value pairs for the environment the lookup is for.