Improve error handling when setting Master Address
This commit pushes the exception handling up from the
PuppetEnterpriseConfig class to the end classes that handle it such as
PuppetEnterpriseStep or PupppetEnterpriseManagement. It's much clearer
what's going on now.

There's also a new validatePuppetMaster method that in
PuppetEnterpriseConfig that should make it easier in the future to do
proper form validation when saving saving the configuration's settings.